requirements.txt
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dotenv==0.9.1
six==1.11.0
Werkzeug==0.14.1
